Adapted CLI to use the new MemoryInterface.

+++ src/lib/stlink.rs
@@ -194,7 +194,7 @@ impl<'a> DAPAccess for STLink<'a> {
     }
 
     /// Writes a value to the DAP register on the specified port and address.
-    fn write_register(&mut self, port: u16, addr: u16, value: u32) -> Result<(), Self::Error> {
+    fn write_register(&mut self, port: u16, addr: u32, value: u32) -> Result<(), Self::Error> {
         if (addr & 0xf0) == 0 || port != Self::DP_PORT {
             let cmd = vec![
                 commands::JTAG_COMMAND,
